fix search-filter: stop propagation with multiple search in the same page
The click stop-propagation was set when clicking on the search icon. It stop the current
click. Thus, with multiple search filter in the same page, we could have multiple ones
opened. By putting the click handler creation in a timeout, it fixes the problem. Now the
first opened filter is closed when opening a new one.