fix search-field: display the cleaning cross when reopening the field
On dropdown or selects, when we opened the dropdown with a value already in the field,
the cleaning cross was not displayed unless we made a change inside it. By adding the
"search-filter--is-active" in the link function if the input already has a value, the bug is fixed.

fix second point of #26