feat selects: converter model=>items & vice versa
Before the selects always returned a list of the object displayed in
the dropdown. By adding two converters ("modelToSelection" &
"selectionToModel"), the user can manage the value returned by the
components. With this, the select can return just an item for
non-multiple or also return just the key (or another property) of a
more complex object. The two converter are function passed to the
directive.

Close #30
BROKEN: selects "selected" is replaced by "model"