fix selects: pointers between model & selected
Before, a simple assignation was down. Thus, the model and the selected
values were the same in case of a multiple select. Because of that,
the callbacks managed for the selected was not called as the model
changed as the same time, waking its watcher before the selected's one.
By using "angular.copy", we isolate the two values and get the expected
behaviour.

fix #30