fix selects: indexof between objects
By isolating the two values of model & selected inside the directive,
a bug was created because the indexOf was not working anymore (the
object were differents). So the object listed in a multiple were never
checked even if they were already selected in the model.
By using a custom indexof based on "angular.equals", the indexOf are
fixed and all the logic previously working works again.

fix #30