fix selects: selected items init in the dropdown
By canceling the first watch over the selected items, they were not
initialized in the DOM anymore inside the dropdown. The cancel of the
watch was added to stop a potential loop between the watchers of model
& selected.
By moving the dropdown's DOM management before the canceling, the DOM is
initialized correctly again. As the DOM update does not touch to the
model, there is no loop risk.

fix #30