feat transclude: add iterative parent option
The "ng-if" and whole other Angular Directives create their own scope.
When using it inside in combination with a "ng-transclude" inside a
custom directives, it might cause some unexpeceted behaviour with the
transcluding content not able to access the right scope without using
multiple "$parent".
By implementing a recursive "ng-transclude", the scope used in the
transcluded content can be selected for each use case simply by
indicating the count of "$parent" to the scope wanted.