fix dropdown: "grand-parent" scope for transclude
Before, we needed to use "$parent.$parent" in the ng-click of the
dropdown link because of the scope created by the "ng-if" and the
"lx-dropdown" directives.
By using the grand-parent scope, we can manage the content of the
"lx-dropdown-menu" as if it was in the same scope of the template we're
currently working on. It's simpler for the users, and the expected
behaviour.