fix text-fields: workaround HTML5 validation model
Before, the model wasn't updated if the value in the text field was not
valid regarding of the HTML5 type (email or date for example). By
watching the "val()" of the angular.element, we are able to still
detect that the value has changed and update the CSS classes to slide
up the label.

fix #37