fix tabs: add timeout on tab links
Before, when clicking on a tab link, the tab was changed before any
click handlers had time to receive the click. By adding a timeout, the
click is managed by all the handlers that have to manage it before the
tab is changed. For example, the dropdown close as expected when
changing a tab.