fix tabs: update heading when scope changes
The heading was set only at the load of the tab by the "addTab" method.
So, if the value of the heading changed, it was not updated. By passing
the scope to the "addTab" method, we can now directly use the latest
scope's value. Also, it removes a use of "ng-bind-html".