fix selects: use child scope inside the directive
Before, if we wanted to use a scope method inside the
"lx-select-selected" or "lx-select-choices", we couldn't because the
directive scope was isolated.
By using a child scope, we keep access to all of the data from the
parent scope. Making this, we now use "ng-model" to manage the model as
the two-ways binding is only doable in an isolated scope.

BROKEN:
selects now use "ng-model" instead of "model"