fix dropdown: Use ng-show instead of ng-if
'ng-if' was not managing the $destroy correctly. If the dropdown was
creating watchers (by using DOM binding), those watchers were not
destroyed on close and was recreated on re-open. Thus, the number of
watchers used was more and more important.

By using ng-show, we avoid this problem. The caveat is that all the
watchers of the dropdown are always active, not only when the dropdown
is open.