fix dropdown: use child scope for directive
Using isolated scope created complicated logic with scope management in
ng-transclude. It also created a bug where the watchers created by using
"{{ }}" in the DOM would not be destroyed on the dropdown close.

By using a child scope, we keep the ability to use elements from the
scope in which the dropdown is used, but we fix the leaking watchers.