fix select: new DOM filter manage filter in params
By using the new filter created inside the lx-select directive, the
DOM filter by text was applied even if we used an external filter (the
filter params of the directive). By managing this external filter inside
the new filter, we can choose to not use the DOM filter when an external
one is defined.