fix dialog: resize event from true size
Before, we handled the resize event by using the current size of the
dropdown. It made the dropdown sticky to the bottom in a slow resize,
even once the dropdown was at the max height of its content.

By resetting the style of the dropdown before doing the calculations,
the dropdown height is always the one needed.