fix tabs: two way data binding with $parse
To simulate the two way data binding without the isolated scope, we used
the $eval ("scope.test = scope.$parent.$eval(attr.test)"). It works only
for object.

Active tab is a primitive type. The $eval way worked to get the value,
but not to set the value. By using a $parse on the attribute, we can
assign the value.