fix dialog: better scrollbar management
The check done every 500ms watched only the current height of the dialog
before updating the height & scrollbar. If the current height was set,
the height did not change.

By checking the real height and not the current, and also checking the
window height, the scrollbar is updated as wanted.

Also, better performances by optimizing the code executed.