fix text field: better height for textarea
The height was removed and then set after each modification of the
textarea value. In a dialog for example, if there was a scrollbar, the
height once removed made the scrollbar go to top.

By creating a fake textarea to find out the height, we avoid removing
the height from the current textarea.