feat notification: add the ability to close with escape/return key
These feature is enabled by default.
When pressing the return key in a confirm or an alert, then it
close and validate the confirm (or simply close the alert)
When pressing escape in a confirm, then it close and cancel the
confirm
Add a new parameter to the confirm and alert functiond: 'unbind'
to disable this feature.