feat eventScheduler: add an event scheduler
This scheduler wiill be used to schedule events management in
LumX, particularly the "Escape to close" feature.
An element can register to the scheduler with an event name
and an handler for the event. It will receive an scheduler unique
identifier.
When there is at least one element registered for an event name,
the scheduler will listen to this event.
When the event triggers, the scheduler executes the handler in
the reversed order of the registrering (the last registered will be
executed first).
If the event is "stopPropagated" in the handler, then the
scheduler doesn't execute any other handler.
An element can then unregister itself from the scheduler by
giving it's unique identifier. It will then be not executed anymore
by the scheduler.
When there is no more element registered to an event, then the
scheduler stop listening to this event.