feat dialog: allow to cancel a dialog closing
It's now possible to pass a new parameter to dialog: 'before-close'.
This must point to a function that will be called just before closing the dialog.

If the function returns true, then the dialog is closed. If it returns false, the dialog is not
closed.
The function can also returns a promise. When the promise resolves, the dialog is closed.
If the promise is rejected (or never resolves), then the dialog is not closed.

This allow for example to ask confirmation to the user to close the dialog
(with LxNotificationService.confirm for example)