feat(notification): always call the callback when the notif closes
Not only when there is an action.
This way, it allow to have an action when a notification closes, either
by clicking on it or when the delay is expired

BROKEN:
Notification now always call the callback with either
true or false. Be sure to check the value sent as the
first parameter of the callback