feat select: add a "panes" choices view mode
This view mode allow to display multiple panes for selecting choices.
The choices must be provided according to a specific data model.
If filter is enable, then it will perform a search through the path of
the panes to search for matching items. Matching items will be
highlighted and containing panes will be auto-opened.

For more information regarding this view mode, see the examples