fix dropdown: make the dropdown scrollable
If the content of the dropdown is to big to fit in the remaining of the
screen (to the top or the bottom according to the position of the target
of the dropdown), then make it scrollable.

Also, improve the recomputation of the height of the dropdown. Watch the
changes of the content instead of using an interval.